requirements.txt
# Core ML
torch>=2.1.0,<2.2.0
torchvision>=0.16.0,<0.17.0

# Scientific computing
numpy>=1.24.0,<1.26.0
scipy>=1.11.0,<1.13.0
scikit-learn>=1.3.0,<1.5.0

# Data handling
pandas>=2.0.0,<2.3.0

# Visualization
matplotlib>=3.7.0,<3.9.0
seaborn>=0.12.0,<0.14.0

# Utilities
tqdm>=4.65.0,<5.0.0
pyyaml>=6.0.0,<7.0.0
termcolor>=2.3.0,<3.0.0

# Computer Vision
opencv-python>=4.8.0,<4.10.0

# Development
jupyter>=1.0.0,<2.0.0
ipython>=8.17.0,<9.0.0

# Phase 2: Report generation
reportlab>=4.0.0,<5.0.0
markdown>=3.0.0,<4.0.0

# Phase 2: Additional (optional)
# adversarial-robustness-toolbox>=1.16.0,<1.17.0
# foolbox>=3.3.0,<3.4.0

# ============================================================================
# DATABASE DEPENDENCIES - PHASE 5.1
# ============================================================================

# PostgreSQL Driver
psycopg2-binary==2.9.9

# SQLAlchemy ORM
SQLAlchemy==2.0.25

# Database Migrations (for schema evolution)
alembic==1.13.1

# Connection Pooling
SQLAlchemy-Utils==0.41.1
